shared_notes/requirements.txt
Django
psycopg2-binary
djangorestframework
# django-rest-swagger
